refactor(messaging): split message_row.dart's action-bar widgets; add missing test coverage
message_row.dart crossed 1000 lines with the voice-chat menu fix in this PR.
Extracts the trailing-edge action-bar widgets (_FloatingActionsOverlay,
_MaybeTappable, _GutterTimestamp, _HoverActions, _MessageActions) into a new
message_row_actions.dart part file, following the existing convention of
splitting message_pane.dart by concern (attachments.dart, composer.dart,
reactions.dart, mute_button.dart).

Also adds test coverage for two code paths this PR introduced without any:
- A failed message delete now shows a "Failed to delete message" snackbar
  and leaves the row in place, exercised end-to-end through MessagePane
  against a mocked 500 response.
- BansApi.create's delete_message_seconds field (the whole point of the
  ban-purge feature) and deleted_message_count parsing, which had zero
  existing coverage for the class.

flutter analyze and the full flutter/dart test suites are unaffected (981 +
154 tests passing).
